test: convert physical-tenant authorization ITs to MultiDbTest
Move the three PT authorization ITs (MultiPhysicalTenantAuthorizationIT,
PhysicalTenantAuthorizationEnablementIT, PhysicalTenantLogicalTenantScopingIT)
to qa/acceptance-tests as @MultiDbTest + @MultiDbPhysicalTenants, gated to the
RDBMS matrix (RDBMS_.*) so they run across the supported RDBMS databases and
skip on ES/OS. They get fresh-storage-per-run isolation from the framework,
retiring the hand-rolled isolation the bespoke PhysicalTenantsITHelper needed.
The helper and its two remaining out-of-scope consumers
(PhysicalTenantIsolationIT, PhysicalTenantRestIsolationIT) are left untouched.
